requirements.txt
torch
numpy
pandas
PyYAML
python-Levenshtein
matplotlib
